requirements.txt
asgiref==3.6.0
Django==3.1.12
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
mongoengine==0.27.0
pymongo==3.12.3
python-decouple==3.8
python-memcached==1.59
pytz==2022.7.1
six==1.16.0
sqlparse==0.2.4
tzdata==2022.7
